Pickup initial force now working correctly

+++ MinecraftClone/MinecraftClone/Pickup.h
@@ -22,7 +22,7 @@ public:
 	//On Player disgard
 	Pickup(eCubeType cubeType, const glm::vec3& position, const glm::vec3& initialVelocity);
 	//On destroyed Cube
-	Pickup(eCubeType cubeType, const glm::ivec3& position);
+	Pickup(eCubeType cubeType, const glm::vec3& position);
 	Pickup(Pickup&&) noexcept;
 	Pickup& operator=(Pickup&&) noexcept;
